requirements.txt
m3u8
requests
retry
pycryptodome